Part9 (#5)
* Exercise 9.1

Write the function calculateBmi and create npm script for running the program with command npm run calculateBmi

* Exercise 9.2

Write a function calculateExercises that calculates the average time of daily exercise hours and compares it to the target amount of daily hours

* Exercise 9.3

Change the previous exercises so that you can give the parameters of bmiCalculator and exerciseCalculator as command-line arguments, and handle exceptions and errors appropriately

* Exercise 9.4

Add Express to your dependencies and create an HTTP GET endpoint hello that answers 'Hello Full Stack!', add start and dev package scripts, replace tsconfig.json

* Exercise 9.5

Add an endpoint for the BMI calculator that can be used by doing an HTTP GET request to endpoint bmi and specifying the input with query string parameters

* Exercise 9.6

Configure your project to use the above eslint settings and fix all the warnings.

* Exercise 9.7

Add an endpoint to your app for the exercise calculator. It should be used by doing an HTTP POST request to endpoint exercises with the input in the request body

* Exercise 9.8

Initialize patientor backend

* Exercise 9.9

Clone the pre-made frontend and ensure the ping request upon startup works correctly

* Exercise 9.10

Create a type Diagnose and use it to create endpoint /api/diagnoses for fetching all diagnoses with HTTP GET.

* Exercise 9.11

Create data type Patient and set up the GET endpoint /api/patients which returns all patients to the frontend, excluding field ssn. (Side note: had a good chuckle at the provided patients' data)

* Exercise 9.12

Create a POST endpoint /api/patients for adding patients

* Exercise 9.13

Set up safe parsing, validation and type guards to the POST /api/patients request.

Refactor the gender field to use an enum type.

* Exercise 9.14

Create a new Create React App with TypeScript

* Exercise 9.15

I ran into so, so, so many ridiculous problems here. For one, had to create an entirely new project and copy-paste existing code. Every config file and line of code exactly the same, but it stopped being utterly broken?

I don't think I'm a fan of TypeScript... I'll just use C# or whatever if I really need static typing

* Exercise 9.16

Create an endpoint /api/patients/:id that returns all of the patient information for one patient

* Exercise 9.17

Create a page for showing a patient's full information in the frontend

* Exercise 9.18

Define action creator functions in the file src/state/reducer.ts and refactor the code to use them (Also fixed bug where currentPatient wasn't resetting - maybe not the most elegant fixed but it'll do for now)

* Exercise 9.19

Define the types OccupationalHealthcareEntry and HospitalEntry so that those conform with the example data. Ensure that your backend returns the entries properly when you go to an individual patient's route

* Exercise 9.20

Extend a patient's page in the frontend to list the date, description and diagnose codes of the patient's entries

* Exercise 9.21

Fetch and add diagnoses to the application state from the /api/diagnoses endpoint. Use the new diagnosis data to show the descriptions for patient's diagnosis codes

* Exercise 9.22

Extend the entry listing in the patient's page to include the Entry's details with a new component that shows the rest of the information of the patient's entries distinguishing different types from each other

* Exercise 9.23

add endpoint /api/patients/:id/entries to your backend, through which you can POST an entry for a patient

* Exercise 9.24

add a form for adding an entry to a patient. This one was NOT easy! Definitely better ways to get the values from DiagnosisSelection (which was not explained until the next exercise...), but it works and I'm very tired in terms of both being sleepy and of TypeScript, so I'll take what I can get.