Organize (#9)
* Add missing parts

These parts were completed in other repos

* Modify Readmes

Removed auto-generated and extraneous readmes and started writing proper project readmes

* Modify Gitignores

* Add project missing from Part7

* Add some hour totals

I would add the rest, but the FSO site appears to be having issues and I can't access my submissions!

* Rename directories

* Add hours

* Pull from main (#8)

* Part12 (#6)

* Initial commit

* Exercises 12.1-12.4 & config

* Exercise 12.5-12.7 & config

* Exercise 12.9

Set up Redis for the project

* Exercise 12.10

Implement a todo counter that saves the number of created todos to Redis

* Exercise 12.11

Use redis-cli to access the db

* Exercise 12.12

Ensure persistence of Redis data

* Exercise 12.13

Todo application frontend containerized

* Exercise 12.14

Utilize multi-stage builds to use a separate build stage for testing

* Exercise 12.15

Use volumes to enable the development of the todo-frontend while it is running inside a container

* Exercise 12.16

Use volumes and Nodemon to enable the development of the todo app backend while it is running inside a container

* Exercise 12.17

Set up an Nginx reverse proxy server in front of todo-frontend

* Exercises 12.18 & 12.19

Connect the services, todo-frontend with todo-backend (Hot-reload of frontend doesn't work due to WSL2, I'm not spending the time to fix that right now)

* Exercise 12.20

Create a production docker-compose.yml with all of the services, Nginx, todo-backend, todo-frontend, MongoDB and Redis

* Exercise 12.21

Containerized dev environment for Bloglist App

* Exercise 12.22

Containerized prod environment for Bloglist App. I had tons of trouble with Nginx and spent way too much time on this - it could definitely be a much more "true" prod environment, and I never got the tests quite working, but I've run out of time and I'm calling this finished. From what I've read, this isn't really the best way to set up a prod environment anyway, as opposed to K8s or Docker Swarm etc.

* Tidy repo and add missing projects (#7)