Exercise 13.7
The instructions say "Centralize the application error handling in middleware as in part 3. You can also enable middleware express-async-errors as we did in part 4". I'm taking this to mean that I'm not supposed to validate the data or anything, just handle express route errors which the express-async-errors library does entirely