Part13 (#10)
* Exercises 13.1 - 13.4

Create a Heroku App initialized with a PostgreSQL db, for a new version of the Bloglist backend. Create GET, POST, and DELETE routes

* Exercises 13.5 & 13.6

Restructure application and implement support for changing the number of a blog's likes

* Exercise 13.7

The instructions say "Centralize the application error handling in middleware as in part 3. You can also enable middleware express-async-errors as we did in part 4". I'm taking this to mean that I'm not supposed to validate the data or anything, just handle express route errors which the express-async-errors library does entirely

* Exercise 13.8

Add support for users to the application

* Exercise 13.9

Implement validation and error handling for usernames

* Exercise 13.10

Expand the application so that the current logged-in user identified by a token is linked to each blog added

* Exercises 13.11 & 13.12

Make deletion of a blog only possible for the user who added the blog, and modify the routes for retrieving all blogs and all users so that each blog shows the user who added it and each user shows the blogs they have added (was already completed)

* Exercise 13.13

Implement filtering by keyword in the application for the route returning all blogs

* Exercise 13.14

Expand the filter to search for a keyword in either the title or author fields

* Exercise 13.15

Modify the blogs route so that it returns blogs based on likes in descending order

* Exercise 13.16

Make a route for the application /api/authors that returns the number of blogs for each author and the total number of likes. Bonus task: order the data returned based on the number of likes

* Exercise 13.17

Make a migration that intializes the database. Add created_at and updated_at timestamps for both tables.

* Exercise 13.18

Expand the application (by migration) so that the blogs have a year written attribute. Also fixed the timestamps from the previous exercise.

* Add rollback

* Exercise 13.19

Give users the ability to add blogs on the system to a reading list

* Exercise 13.20

Add functionality to the application to support the reading list

* Update README.md

* Exercise 13.21

Expand the single-user route so that each blog in the reading list shows also whether the blog has been read and the id of the corresponding connection table row (was already done, just cleaned it up a bit)

* Exercise 13.22

Implement functionality in the application to mark a blog in the reading list as read. Also refactored and simplified the db schema

* Exercise 13.23

Modify the route that returns a single user's information so that the request can control which of the blogs in the reading list are returned

* Exercise 13.24

The grand finale!!

Expand the system so that the user who has lost access will not be able to perform any actions that require login

* Update readmes